refactor: move getUserId() from startApexDebugLogging.ts to authInfo.ts (better place for a utility function)
