+++ packages/salesforcedx-vscode-core/src/commands/stopApexDebugLogging.ts
@@ -10,8 +10,8 @@ import * as vscode from 'vscode';
 import { TRACE_FLAG_EXPIRATION_KEY } from '../constants';
 import { WorkspaceContext } from '../context';
 import { disposeTraceFlagExpiration } from '../decorators/traceflagTimeDecorator';
+import { OrgAuthInfo } from '../util';
 import { handleStartCommand, handleFinishCommand } from '../utils/channelUtils';
-import { getUserId } from './startApexDebugLogging';
 
 const command = 'stop_apex_debug_logging';
 
@@ -22,7 +22,7 @@ export const turnOffLogging = async (extensionContext: vscode.ExtensionContext):
 
   // Check if a TraceFlag already exists for the current user
   const traceFlags = await connection.tooling.query(
-    `SELECT Id, ExpirationDate FROM TraceFlag WHERE LogType = 'DEVELOPER_LOG' AND TracedEntityId = '${await getUserId(connection)}'`
+    `SELECT Id, ExpirationDate FROM TraceFlag WHERE LogType = 'DEVELOPER_LOG' AND TracedEntityId = '${await OrgAuthInfo.getUserId()}'`
   );
   const traceFlagExists = traceFlags.records.length > 0;
